+++ delivery/src/main/java/com/nowayback/delivery/presentation/delivery/dto/request/UpdateDeliveryRecipientInfoRequest.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.presentation.dto.request;
+package com.nowayback.delivery.presentation.delivery.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
