+++ delivery/src/main/java/com/nowayback/delivery/presentation/delivery/dto/request/UpdateDeliveryStatusRequest.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.presentation.dto.request;
+package com.nowayback.delivery.presentation.delivery.dto.request;
 
 import com.nowayback.delivery.domain.delivery.vo.DeliveryStatus;
 import jakarta.validation.constraints.NotNull;
